Add Newline/Indentation Tokens
Add to the TokenLexer the ability to recognize levels of indentation from
line to line, and parse the combination of newlines and indentation into
tokens. Later split these, as the grouping was mainly for associating the
two.

This temporarily breaks the AST parser, but will be helpful in adding
delimiting of expressions, optionally with a delimiter character such as a
semicolon.