Fold in TokenLexer Post-Processing, Fixes/Tests for ASTParser
Get rid of the LexerToParser, as it wasn't being used in tests, and make
it simply a step of the tokens Parser in the TokenLexer (the top level
parser used by that one). Adjust tests accordingly.

Ensure that the end of input is seen as equivalent to a delimiter, for the
purposes of expressions (as there is no guarantee a newline will exist
after the last definition's expression).