Structure the Compiler For Incremental Compilation
Use Dotty's example and build this a little more in a database-oriented
direction, by having
- the Compiler get made up of phases
- each of those phases can be composed with each other to make combined
  phases, if their input/output match
- each of those phases output is recorded (composed phases have each
  subphase's output recorded)

This opens the door to the Compiler running as a server rather than just a
single-run pass and/or reusing output from previous runs if matching
input is found.

The unit of compilation is still only a single string, so in future work,
projects with multiple files/parts should be definable such that
individual units of compilation can have their output cached.